+++ ThirdPartyAdapters/line/CHANGELOG.md
@@ -1,8 +1,16 @@
 ## LINE Android Mediation Adapter Changelog
 
+#### Version 3.0.1.1 (In progress)
+- Native ad impressions are now overwritten by FiveAd SDK when using Google
+Mobile Ads SDK v24.4.0 or above.
+
+Built and tested with:
+- Google Mobile Ads SDK version 24.9.0.
+- FiveAd SDK version 3.0.1.0.
+
 #### Version 3.0.1.0
-- Mapped Google Mobile Ads SDK underage flags to Line SDK child treatment
-methods.
+- Added support for forwarding the tagForUnderAgeOfConsent Google Mobile Ads SDK
+parameter to the FiveAd SDK.
 - Verified compatibility with FiveAd SDK version 3.0.1.0.
 
 Built and tested with: