+++ ThirdPartyAdapters/pubmatic/CHANGELOG.md
@@ -1,8 +1,15 @@
 ## PubMatic Android Mediation Adapter Changelog
 
+#### Version 4.12.0.0
+- Verified compatibility with Pubmatic SDK 4.12.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 24.9.0.
+- Pubmatic SDK version 4.12.0
+
 #### Version 4.11.0.1
-- Mapped Google Mobile Ads SDK underage flags to Pubmatic OpenWrap SDK COPPA
-method.
+- Added support for forwarding the `tagForUnderAgeOfConsent` Google Mobile Ads SDK
+parameter to the PubMatic SDK.
 
 Built and tested with:
 - Google Mobile Ads SDK version 24.9.0.